requirements.txt
requests>=2.26.0
pyyaml>=5.4.1 